requirements.txt
streamlit==1.48.0
pillow==10.3.0